Added schedule to tour
