+++ frontend/pages/admin/data/tours/index.tsx
@@ -7,9 +7,9 @@ import { useRouter } from "next/router";
 import MaterialReactTable, { MRT_ColumnDef } from "material-react-table";
 import { Box, IconButton, Tooltip } from "@mui/material";
 import { Button } from "react-bootstrap";
-import { Delete, Edit } from "@mui/icons-material";
+import {CalendarMonth, Delete, Edit} from "@mui/icons-material";
 import { BuildingInterface, getAddress } from "@/lib/building";
-import { TourView } from "@/types";
+import {TourView} from "@/types";
 import { TourDeleteModal } from "@/components/admin/tourDeleteModal";
 
 // https://www.figma.com/proto/9yLULhNn8b8SlsWlOnRSpm/SeLab2-mockup?node-id=68-429&scaling=contain&page-id=0%3A1&starting-point-node-id=118%3A1486
@@ -131,6 +131,13 @@ function AdminDataTours() {
         setSelectedTour(null);
     }
 
+    async function routeToGarbageSchedule(tourView: TourView) {
+        await router.push({
+            pathname: `/admin/data/garbage-collection`,
+            query: { tour: tourView.tour_id },
+        });
+    }
+
     return (
         <>
             <AdminHeader />
@@ -181,6 +188,16 @@ function AdminDataTours() {
                                 }}
                             >
                                 <Delete />
+                            </IconButton>
+                        </Tooltip>
+                        <Tooltip arrow placement="right" title="Vuilophaling">
+                            <IconButton
+                                onClick={() => {
+                                    const tourView: TourView = row.original;
+                                    routeToGarbageSchedule(tourView).then();
+                                }}
+                            >
+                                <CalendarMonth />
                             </IconButton>
                         </Tooltip>
                     </Box>
